backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.11.0
python-dotenv==1.0.1
